Removes size_t from shareable types to prevent 32/64-bit errors. Adds unsigned int to shareable types. In the future, use int32/player_id_t for player IDs and uint32/entity_id_t for entity IDs.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@11189 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasUI/ScenarioEditor/Tools/Common/ObjectSettings.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2009 Wildfire Games.
+/* Copyright (C) 2012 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -36,7 +36,7 @@ class ObjectSettings
 public:
 	ObjectSettings(Observable<std::vector<AtlasMessage::ObjectID> >& selectedObjects, int view);
 
-	size_t GetPlayerID() const;
+	int GetPlayerID() const;
 	void SetPlayerID(int playerID);
 
 	void SetView(int view);
@@ -61,7 +61,7 @@ private:
 	int m_View;
 
 	// 0 = gaia, 1..inf = normal players
-	size_t m_PlayerID;
+	int m_PlayerID;
 
 	// Set of user-chosen actor selections, potentially a superset of any single
 	// actor's possible variants (since it doesn't get reset if you select
